requirements.txt
# JSBSim XML Generator - Python Dependencies
# ============================================
#
# This tool supports 2 input methods:
#   1. FMS→JSBSim conversion (src/fms_to_jsbsim/)
#   2. Excel→JSBSim conversion (src/excel_to_jsbsim/)
#
# Installation:
#   pip install -r requirements.txt
#
# Verified Environment:
#   Python 3.8+
#   Tested on: Python 3.13

# ===== Common Dependencies (Both Methods) =====

# Configuration file management
PyYAML>=5.4.0

# ===== Excel→JSBSim Conversion =====

# Excel file I/O
openpyxl>=3.0.0

# Data manipulation
pandas>=1.3.0
numpy>=1.21.0

# Numerical computation (for trim/stability calculations)
scipy>=1.7.0

# ===== FMS→JSBSim Conversion =====
# (Uses standard library only: json, re, pathlib, typing, datetime)
# No additional dependencies required

# ===== JSBSim (Required for Both Methods) =====

# JSBSim flight dynamics engine
# Note: JSBSim is LGPL 2.1+ licensed. This project uses it as an external library.
# Users must comply with JSBSim's LGPL license terms.
jsbsim>=1.1.0

# Optional: Google Sheets API (if using Google Sheets instead of Excel)
# Uncomment below if you want to use Google Sheets integration
# gspread>=5.0.0
# google-auth>=2.0.0
